Add files via upload
(note-- Final Code just copy and paste in Arduino IDE.)
The Fingerprint Based Voting System is one of the voting techniques in which people can cast
their votes to their candidates. These days voting machine has become effective toll compared
with traditional way of voting using paper. Thus, we decided to design the machine to overcome
the already existing voting technique. The main objective of this is to give security to the
valuable votes of the voters and make voting process easy as much as possible.
In this, Arduino is the main part and other devices are connected to it. The fingerprint module
is connected to Arduino which stores the data for voters for verification. LCD is used to display
the status of the voters. Firstly, the voter must give thumb impression on the module to start the
process of voting. The Arduino will check the data and LCD will display the output. Fingerprint
module plays an important role in this system as it secures voter’s data.